Refactor: rename banner component for clarity
