+++ src/components/player/SignedOutBanner.tsx
